requirements.txt
# Flask-Tesseract5
image
pillow==7.1.2
pytesseract
scikit-build
opencv-python
Flask
flask-restful

# Voon
youtube-dl
nltk
PyGithub
requests